Move test files to match src files
